Use specific path to access resources
There are different ways to run a Python project, some of which won't
work properly when using relative paths. This change makes the
assumption that the resources directory is in the same directory with
universe.py. There are, perhaps, better ways to do this, but without
knowing more about the overall direction of the project, it's hard to
propose something. This change is more robust and will likely suffice
for the foreseeable future.