Add a config file for development
Added config_dev.ini to the project, which allows a developer to toggle testing_mode, set the starting map, and starting map position. I plan to add more features to this config when they become important.

In game.py, added support for the config file by importing configparser and ast. ast is for interpreting a tuple from a string, which was needed for the starting location.